requirements.txt
pandas
selenium
webdriver-manager
requests
python-dotenv
